+++ raw/raw_schema-id.go
@@ -82,20 +82,20 @@ func (s SchemaID) Equal(id SchemaID) bool {
 	return s.Group == id.Group && s.Kind == id.Kind
 }
 
-func (s SchemaID) InputTopic(branch base.Branch) libkafka.Topic {
-	return BuildTopic(s, branch, "input")
+func (s SchemaID) InputTopic(prefix base.TopicPrefix) libkafka.Topic {
+	return BuildTopic(s, prefix, "input")
 }
 
-func (s SchemaID) EventTopic(branch base.Branch) libkafka.Topic {
-	return BuildTopic(s, branch, "event")
+func (s SchemaID) EventTopic(prefix base.TopicPrefix) libkafka.Topic {
+	return BuildTopic(s, prefix, "event")
 }
 
-func (s SchemaID) ResultTopic(branch base.Branch) libkafka.Topic {
-	return BuildTopic(s, branch, "result")
+func (s SchemaID) ResultTopic(prefix base.TopicPrefix) libkafka.Topic {
+	return BuildTopic(s, prefix, "result")
 }
 
-func (s SchemaID) CommandTopic(branch base.Branch) libkafka.Topic {
-	return BuildTopic(s, branch, "request")
+func (s SchemaID) CommandTopic(prefix base.TopicPrefix) libkafka.Topic {
+	return BuildTopic(s, prefix, "request")
 }
 
 var validateGroupRegex = regexp.MustCompile(`^[a-z][a-z0-9]*$`)